Backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.2
scikit-learn==1.4.2
google-generativeai==0.7.2
python-dotenv==1.0.1
python-multipart==0.0.9
streamlit
